Moved codec tests into test_code.py
